Update group dao and interface to only do shallow deletes (unless requested specifically)

+++ src/main/java/uk/ac/cam/cl/dtg/segue/dao/users/PgUserGroupPersistenceManager.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright 2014 Stephen Cummins
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -24,15 +24,19 @@ import java.sql.Timestamp;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import com.google.api.client.util.Sets;
+import com.google.common.collect.Maps;
 import org.apache.commons.lang3.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import uk.ac.cam.cl.dtg.segue.dao.SegueDatabaseException;
 import uk.ac.cam.cl.dtg.segue.database.PostgresSqlDb;
+import uk.ac.cam.cl.dtg.segue.dos.GroupMembership;
+import uk.ac.cam.cl.dtg.segue.dos.GroupMembershipStatus;
 import uk.ac.cam.cl.dtg.segue.dos.GroupStatus;
 import uk.ac.cam.cl.dtg.segue.dos.UserGroup;
 import com.google.api.client.util.Lists;
@@ -65,8 +69,8 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
             PreparedStatement pst;
             pst = conn
                     .prepareStatement(
-                            "INSERT INTO groups(group_name, owner_id, created, last_updated)"
-                            + " VALUES (?, ?, ?, ?);",
+                            "INSERT INTO groups(group_name, owner_id, group_status, created, last_updated)"
+                            + " VALUES (?, ?, ?, ?, ?);",
                             Statement.RETURN_GENERATED_KEYS);
             pst.setString(1, group.getGroupName());
             pst.setLong(2, group.getOwnerId());
@@ -77,8 +81,11 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
             } else {
                 created = new Timestamp(new Date().getTime());
             }
-            pst.setTimestamp(3, created);
+
+            pst.setString(3, GroupStatus.ACTIVE.name());
+
             pst.setTimestamp(4, created);
+            pst.setTimestamp(5, created);
 
             if (pst.executeUpdate() == 0) {
                 throw new SegueDatabaseException("Unable to save group.");
@@ -108,13 +115,15 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
         PreparedStatement pst;
         try (Connection conn = database.getDatabaseConnection()) {
             pst = conn.prepareStatement(
-                    "UPDATE groups SET group_name=?, owner_id=?, created=?, archived=?, last_updated=? WHERE id = ?;");
+                    "UPDATE groups SET group_name=?, owner_id=?, created=?, archived=?, group_status=?, last_updated=? WHERE id = ?;");
+
             pst.setString(1, group.getGroupName());
             pst.setLong(2, group.getOwnerId());
             pst.setTimestamp(3, new Timestamp(group.getCreated().getTime()));
             pst.setBoolean(4, group.isArchived());
-            pst.setTimestamp(5, new Timestamp(group.getLastUpdated().getTime()));
-            pst.setLong(6, group.getId());
+            pst.setString(5, group.getStatus().name());
+            pst.setTimestamp(6, new Timestamp(group.getLastUpdated().getTime()));
+            pst.setLong(7, group.getId());
             
             log.debug(pst.toString());
             
@@ -122,7 +131,7 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
                 throw new SegueDatabaseException("Unable to save group.");
             }
 
-            return this.findById(group.getId());
+            return this.findGroupById(group.getId());
         } catch (SQLException e) {
             throw new SegueDatabaseException("Postgres exception", e);
         }
@@ -130,16 +139,25 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
 
     @Override
     public void addUserToGroup(final Long userId, final Long groupId) throws SegueDatabaseException {
+        // first check if they already have a membership record
+        if (this.hasMembershipForGroup(userId, groupId)) {
+            this.setUsersGroupMembershipStatus(userId,groupId, GroupMembershipStatus.ACTIVE);
+            return;
+        }
+
         try (Connection conn = database.getDatabaseConnection()) {
             PreparedStatement pst;
+
             pst = conn
                     .prepareStatement(
-                            "INSERT INTO group_memberships(group_id, user_id, created) VALUES (?, ?, ?);",
+                            "INSERT INTO group_memberships(group_id, user_id, status, created, updated) VALUES (?, ?, ?, ?, ?);",
                             Statement.RETURN_GENERATED_KEYS);
             pst.setLong(1, groupId);
             pst.setLong(2, userId);
-            pst.setTimestamp(3, new Timestamp(new Date().getTime()));
-            
+            pst.setString(3, GroupMembershipStatus.ACTIVE.name());
+            pst.setTimestamp(4, new Timestamp(new Date().getTime()));
+            pst.setTimestamp(5, new Timestamp(new Date().getTime()));
+
             int affectedRows = pst.executeUpdate();
 
             if (affectedRows == 0) {
@@ -152,17 +170,29 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
     }
 
     @Override
-    public void removeUserFromGroup(final Long userId, final Long groupId) throws SegueDatabaseException {
+    public void setUsersGroupMembershipStatus(final Long userId, final Long groupId, final GroupMembershipStatus newStatus) throws SegueDatabaseException {
         try (Connection conn = database.getDatabaseConnection()) {
             PreparedStatement pst;
-            pst = conn.prepareStatement("DELETE FROM group_memberships WHERE group_id = ? AND user_id = ?");
-            pst.setLong(1, groupId);
-            pst.setLong(2, userId);
-            
-            pst.execute();
+            pst = conn
+                    .prepareStatement("UPDATE group_memberships SET status=?,updated=? WHERE user_id = ? AND group_id = ?");
+            pst.setString(1, newStatus.name());
+            pst.setTimestamp(2, new Timestamp(new Date().getTime()));
+            pst.setLong(3, userId);
+            pst.setLong(4, groupId);
+
+
+            if (pst.executeUpdate() == 0) {
+                throw new SegueDatabaseException("Unable to update membership status.");
+            }
+
         } catch (SQLException e) {
             throw new SegueDatabaseException("Postgres exception", e);
         }
+    }
+
+    @Override
+    public void removeUserFromGroup(final Long userId, final Long groupId) throws SegueDatabaseException {
+        this.setUsersGroupMembershipStatus(userId, groupId, GroupMembershipStatus.DELETED);
     }
 
 
@@ -177,6 +207,8 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
         if (archivedGroupsOnly != null) {
             pstString = pstString +  " AND archived = ?";
         }
+
+        pstString = pstString + " AND group_status <> '" + GroupStatus.DELETED.name() + "'";
 
         return getGroupsBySQLPst(pstString, ownerUserId, archivedGroupsOnly);
 
@@ -201,7 +233,11 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
     public Long getGroupCount() throws SegueDatabaseException {
         try (Connection conn = database.getDatabaseConnection()) {
             PreparedStatement pst;
-            pst = conn.prepareStatement("SELECT COUNT(1) AS TOTAL FROM groups");
+            // we don't want to count 'deleted' groups.
+            pst = conn.prepareStatement("SELECT COUNT(1) AS TOTAL FROM groups " +
+                    "WHERE status <> ?");
+
+            pst.setString(1, GroupStatus.DELETED.name());
 
             ResultSet results = pst.executeQuery();
             results.next();
@@ -212,10 +248,21 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
     }
 
     @Override
-    public UserGroup findById(final Long groupId) throws SegueDatabaseException {
+    public UserGroup findGroupById(final Long groupId) throws SegueDatabaseException {
+        return this.findGroupById(groupId, false);
+    }
+
+    @Override
+    public UserGroup findGroupById(final Long groupId, boolean includeDeletedGroups) throws SegueDatabaseException {
         try (Connection conn = database.getDatabaseConnection()) {
             PreparedStatement pst;
-            pst = conn.prepareStatement("SELECT * FROM groups WHERE id = ?");
+            if(includeDeletedGroups) {
+                pst = conn.prepareStatement("SELECT * FROM groups WHERE id = ?");
+            } else {
+                pst = conn.prepareStatement("SELECT * FROM groups WHERE id = ? AND group_status <> ?");
+                pst.setString(2, GroupStatus.DELETED.name());
+            }
+
             pst.setLong(1, groupId);
 
             ResultSet results = pst.executeQuery();
@@ -237,44 +284,106 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
 
     @Override
     public void deleteGroup(final Long groupId) throws SegueDatabaseException {
+        this.deleteGroup(groupId, true);
+    }
+
+    @Override
+    public void deleteGroup(final Long groupId, boolean markAsDeleted) throws SegueDatabaseException {
         if (null == groupId) {
             throw new SegueDatabaseException("Unable to locate the group requested to delete.");
         }
 
         try (Connection conn = database.getDatabaseConnection()) {
-            try {
                 PreparedStatement pst;
-                pst = conn.prepareStatement("DELETE FROM groups WHERE id = ?");
-                pst.setLong(1, groupId);
-                pst.execute();
-                                
-            } catch (SQLException e) {
-             
-                throw e;
-            }
+                if (markAsDeleted) {
+                        pst = conn
+                                .prepareStatement("UPDATE groups SET group_status=? WHERE id = ?;");
+                        pst.setString(1, GroupStatus.DELETED.name());
+                        pst.setLong(2, groupId);
+
+                        if (pst.executeUpdate() == 0) {
+                            throw new SegueDatabaseException("Unable to mark group as deleted.");
+                        }
+                } else {
+                    pst = conn.prepareStatement("DELETE FROM groups WHERE id = ?");
+                    pst.setLong(1, groupId);
+                    pst.execute();
+                }
+
         } catch (SQLException e1) {
             throw new SegueDatabaseException("Postgres exception", e1);
-        } 
+        }
     }
 
     @Override
     public List<Long> getGroupMemberIds(final Long groupId) throws SegueDatabaseException {
         try (Connection conn = database.getDatabaseConnection()) {
             PreparedStatement pst;
-            pst = conn.prepareStatement("SELECT * FROM group_memberships WHERE group_id = ?");
+
+            pst = conn.prepareStatement("SELECT * FROM group_memberships " +
+                    "WHERE group_id = ? AND status <> ?");
+
             pst.setLong(1, groupId);
+            pst.setString(2, GroupMembershipStatus.DELETED.name());
 
             ResultSet results = pst.executeQuery();
-            
+
             List<Long> listOfResults = Lists.newArrayList();
             while (results.next()) {
                 listOfResults.add(results.getLong("user_id"));
             }
-            
+
             return listOfResults;
         } catch (SQLException e) {
             throw new SegueDatabaseException("Postgres exception", e);
         }
+    }
+
+    @Override
+    public Map<Long, GroupMembership> getGroupMembershipMap(final Long groupId) throws SegueDatabaseException {
+        try (Connection conn = database.getDatabaseConnection()) {
+            PreparedStatement pst;
+
+            pst = conn.prepareStatement("SELECT * FROM group_memberships " +
+                    "WHERE group_id = ? AND status <> ?");
+
+            pst.setLong(1, groupId);
+            pst.setString(2, GroupMembershipStatus.DELETED.name());
+
+            ResultSet results = pst.executeQuery();
+
+            Map<Long, GroupMembership> mapOfResults = Maps.newHashMap();
+            while (results.next()) {
+                mapOfResults.put(results.getLong("user_id"), this.buildMembershipRecord(results));
+            }
+
+            return mapOfResults;
+        } catch (SQLException e) {
+            throw new SegueDatabaseException("Postgres exception", e);
+        }
+    }
+
+    /**
+     * Will look for shallow deleted group membership
+     * @param groupId group to check
+     * @param userId user to check
+     * @return true if they ever had a recorded membership entry in the db.
+     * @throws SegueDatabaseException - if there is db error.
+     */
+    private boolean hasMembershipForGroup(final Long groupId, final Long userId) throws SegueDatabaseException {
+        try (Connection conn = database.getDatabaseConnection()) {
+            PreparedStatement pst;
+
+            pst = conn.prepareStatement("SELECT COUNT(1) AS TOTAL FROM groups WHERE group_id = ? AND user_id = ?");
+            pst.setLong(1, groupId);
+            pst.setLong(2, userId);
+            ResultSet results = pst.executeQuery();
+            results.next();
+            return results.getLong("TOTAL") == 1;
+        } catch (SQLException e) {
+            throw new SegueDatabaseException("Postgres exception: Unable to count groups", e);
+        }
+
     }
 
     @Override
@@ -284,9 +393,9 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
             pst = conn
                     .prepareStatement("SELECT * FROM groups INNER JOIN group_memberships"
                             + " ON groups.id = group_memberships.group_id"
-                            + " WHERE group_memberships.user_id = ?");
+                            + " WHERE user_id = ? AND status <> ?");
             pst.setLong(1, userId);
-
+            pst.setString(2, GroupMembershipStatus.DELETED.name());
             ResultSet results = pst.executeQuery();
             
             List<UserGroup> listOfResults = Lists.newArrayList();
@@ -362,16 +471,19 @@ public class PgUserGroupPersistenceManager implements IUserGroupPersistenceManag
     /**
      * buildGroup. Convenience method to build a group.
      * 
-     * @param set
+     * @param set - the result set
      * @return a Group object
      * @throws SQLException
      *             - if we cannot extract a required property from the results set.
      */
     private UserGroup buildGroup(final ResultSet set) throws SQLException {
-
         return new UserGroup(set.getLong("id"), set.getString("group_name"), set.getLong("owner_id"),
                 GroupStatus.valueOf(set.getString("status")), set.getDate("created"),
                 set.getBoolean("archived"), set.getDate("last_updated"));
+    }
+
+    private GroupMembership buildMembershipRecord(final ResultSet set) throws SQLException {
+        return new GroupMembership(set.getLong("group_id"), set.getLong("user_id"), GroupMembershipStatus.valueOf(set.getString("status")), set.getDate("created"), set.getDate("updated"));
     }
 
     private List<UserGroup> getGroupsBySQLPst(final String pstString, final Long userId, @Nullable final Boolean archivedGroupsOnly) throws SegueDatabaseException {
